requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5

openai==2.26.0
python-dotenv==1.2.2

sentence-transformers==2.7.0
transformers==4.41.2
torch==2.2.2
numpy==1.26.4
scikit-learn==1.8.0

faiss-cpu==1.13.2

requests==2.32.5